new winter terrain...whaddaya think?  off to work...
didn't go too far outside the box...will confer with the programmers to
see if we can implement a more three dimensional path
system...overpasses and such on the next two terrains...will make a snow
covered tree or two in zbrush to populate some of the terrain...should
project nicely on a simple cone...think heavy snow, not lightly
dusted...we can do a simple retexture of a couple of the other assets
and that should do it for this particular terrain...can we have
snowflakes or something drifting across the screen...too busy?  or cool?
(no pun intended)  then similarly, sparks in the volcanic wasteland?  to
shiny? or, totally ON FIRE!!! (completely intended)